requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
